Move test to RetriesTests
